Feat: use time crate instead of chrono (behind feature flag)
The chrono crate has not fixed a CVE for months now, while the
time crate has. chrono is also using an outdated version of
time and this updates it (0.1 -> 0.3).

For people that use `cargo-deny` or that prefer time, offer the option
to use it as the date backend for the `WMIDateTime` type.